ajuste do tipo de dados Float, e sistema de compra de moedas já está funcionando

+++ digital.api/src/digital.domain/Models/Cotacao.cs
@@ -16,7 +16,7 @@ namespace digital.domain.Models
 
         [BsonElement("valorCotado")]
         [BsonRequired]
-        public decimal ValorCotado { get; set; }
+        public double ValorCotado { get; set; }
 
         [BsonElement("dataCotacao")]
         public DateTime DataCotacao { get; set; }
